improved templates and utr computations

+++ src/types.jl
@@ -11,4 +11,4 @@ abstract type FileCollection end
 abstract type InteractionContainer end
 
 LongDNASeqPair = Tuple{LongDNASeq, LongDNASeq}
-CoverageValues = Dict{String, Vector{Float32}}
+CoverageValues = Dict{String, Vector{Float64}}
